alc: widen the transmit TSI to the full field width RFC 3451 defines
AlcPacket's transmit constructor took a uint16_t TSI while Transmitter itself
stores a uint64_t, silently truncating any TSI above 65535 on every packet.
Widened the constructor to uint64_t and now sets tsi_flag (the LCT header's
S flag, adding a second 32-bit TSI word per the S and H flag
mechanism of RFC 3451 clause 5, the LCT generation this library implements) whenever the value exceeds 16 bits, giving the full 48-bit field width
this wire scheme supports. Values that fit in 16 bits keep the original
single-half-word encoding unchanged.

Dropped this commit's original FLUTE-version-2 change: RFC 6726 SS3.1/3.4.1
does require version 2, but 3GPP TS 26.517 cl.6.2.1 -> TS 26.346 Annex L
mandates the FLUTE v1 profile (RFC 3926) for MBMS/5G-MAG applications, and
this library's primary consumers are exactly those applications -- per
davidjwbbc's review on #62. Kept on a separate future/flute-v2-support
branch in case genuine (non-3GPP) FLUTE v2 support is wanted later.

+++ include/AlcPacket.h
@@ -44,7 +44,7 @@ namespace LibFlute {
       *  @param max_size Maximum payload size
       *  @param fdt_instance_id FDT instance ID (only relevant for FDT with TOI=0)
       */
-      AlcPacket(uint16_t tsi, uint16_t toi, FecOti fec_oti, const std::vector<EncodingSymbol>& symbols, size_t max_size, uint32_t fdt_instance_id);
+      AlcPacket(uint64_t tsi, uint16_t toi, FecOti fec_oti, const std::vector<EncodingSymbol>& symbols, size_t max_size, uint32_t fdt_instance_id);
 
      /**
       *  Default destructor.
